@@ -1,16 +1,17 @@
 name: firestore_ui_example
 description: Demonstrates how to use the firestore_ui plugin.
 version: 1.0.0+1
+publish_to: none
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   firestore_ui:
     path: ../
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
