@@ -1,22 +1,23 @@
 name: firestore_ui
 description: A port of firebase_database's FirebaseAnimatedList that works with cloud_firestore, with additional variants
-version: 1.13.0
+version: 1.20.0
 homepage: https://github.com/gildaswise/firestore_ui
 # author: Gildásio Filho <gildasiogx@gmail.com>
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.1.6 
-  firebase_core: ^0.7.0
-  firebase_core_platform_interface: ^3.0.1
-  cloud_firestore: ^0.16.0
-  cloud_firestore_platform_interface: ^3.0.1
-  flutter_staggered_grid_view: ^0.3.2
-  collection: ^1.14.13
+  meta: ^1.3.0
+  firebase_core: ^1.1.0
+  firebase_core_platform_interface: ^4.0.0
+  cloud_firestore: ^1.0.7
+  cloud_firestore_platform_interface: ^4.0.3
+  flutter_staggered_grid_view: ^0.4.0
+  collection: ^1.15.0
 
 dev_dependencies:
   flutter_test:
