@@ -12,7 +12,7 @@ import 'firestore_list.dart';
 
 typedef Widget FirestoreAnimatedListItemBuilder(
   BuildContext context,
-  DocumentSnapshot snapshot,
+  DocumentSnapshot? snapshot,
   Animation<double> animation,
   int index,
 );
@@ -21,9 +21,9 @@ typedef Widget FirestoreAnimatedListItemBuilder(
 class FirestoreAnimatedList extends StatefulWidget {
   /// Creates a scrolling container that animates items when they are inserted or removed.
   FirestoreAnimatedList({
-    Key key,
-    @required this.query,
-    @required this.itemBuilder,
+    Key? key,
+    required this.query,
+    required this.itemBuilder,
     this.onLoaded,
     this.filter,
     this.defaultChild,
@@ -39,20 +39,17 @@ class FirestoreAnimatedList extends StatefulWidget {
     this.shrinkWrap = false,
     this.padding,
     this.duration = const Duration(milliseconds: 300),
-  }) : super(key: key) {
-    assert(query != null);
-    assert(itemBuilder != null);
-  }
+  }) : super(key: key);
 
   /// A Firestore query to use to populate the animated list
   final Query query;
 
   /// Method that gets called once the stream updates with a new QuerySnapshot
-  final Function(QuerySnapshot) onLoaded;
+  final Function(QuerySnapshot)? onLoaded;
 
   /// Called before any operation with a DocumentSnapshot;
   /// If it returns `true`, then dismisses that DocumentSnapshot from the list
-  final FilterCallback filter;
+  final FilterCallback? filter;
 
   /// This will change `onDocumentAdded` call to `.add` instead of `.insert`,
   /// which might help if your query doesn't care about order changes
@@ -60,15 +57,15 @@ class FirestoreAnimatedList extends StatefulWidget {
 
   /// A widget to display while the query is loading. Defaults to a
   /// centered [CircularProgressIndicator];
-  final Widget defaultChild;
+  final Widget? defaultChild;
 
   /// A widget to display if an error ocurred. Defaults to a
   /// centered [Icon] with `Icons.error` and the error itsef;
-  final Widget errorChild;
+  final Widget? errorChild;
 
   /// A widget to display if the query returns empty. Defaults to a
   /// `Container()`;
-  final Widget emptyChild;
+  final Widget? emptyChild;
 
   /// Called, as needed, to build list item widgets.
   ///
@@ -108,7 +105,7 @@ class FirestoreAnimatedList extends StatefulWidget {
   /// view is scrolled.
   ///
   /// Must be null if [primary] is true.
-  final ScrollController controller;
+  final ScrollController? controller;
 
   /// Whether this is the primary scroll view associated with the parent
   /// [PrimaryScrollController].
@@ -118,7 +115,7 @@ class FirestoreAnimatedList extends StatefulWidget {
   ///
   /// Defaults to true when [scrollDirection] is [Axis.vertical] and
   /// [controller] is null.
-  final bool primary;
+  final bool? primary;
 
   /// How the scroll view should respond to user input.
   ///
@@ -126,7 +123,7 @@ class FirestoreAnimatedList extends StatefulWidget {
   /// user stops dragging the scroll view.
   ///
   /// Defaults to matching platform conventions.
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   /// Whether the extent of the scroll view in the [scrollDirection] should be
   /// determined by the contents being viewed.
@@ -145,7 +142,7 @@ class FirestoreAnimatedList extends StatefulWidget {
   final bool shrinkWrap;
 
   /// The amount of space by which to inset the children.
-  final EdgeInsets padding;
+  final EdgeInsets? padding;
 
   /// The duration of the insert and remove animation.
   ///
@@ -159,8 +156,8 @@ class FirestoreAnimatedList extends StatefulWidget {
 class FirestoreAnimatedListState extends State<FirestoreAnimatedList> {
   final GlobalKey<AnimatedListState> _animatedListKey =
       GlobalKey<AnimatedListState>();
-  FirestoreList _model;
-  Exception _error;
+  FirestoreList? _model;
+  Exception? _error;
   bool _loaded = false;
 
   /// Should only be called without setState, inside @override methods here
@@ -196,18 +193,18 @@ class FirestoreAnimatedListState extends State<FirestoreAnimatedList> {
   @override
   void dispose() {
     // Cancel the Firebase stream subscriptions
-    _model.clear();
+    _model!.clear();
     super.dispose();
   }
 
-   void _onError(Exception exception) {
+  void _onError(Exception exception) {
     if (mounted) {
       setState(() {
         _error = exception;
       });
     }
   }
-  
+
   void _onDocumentAdded(int index, DocumentSnapshot snapshot) {
     // if (!_loaded) {
     //   return; // AnimatedList is not created yet
@@ -220,13 +217,13 @@ class FirestoreAnimatedListState extends State<FirestoreAnimatedList> {
         });
       }
     } catch (error) {
-      _model.log("Failed to run onDocumentAdded");
+      _model!.log("Failed to run onDocumentAdded");
     }
   }
 
   void _onDocumentRemoved(int index, DocumentSnapshot snapshot) {
     // The child should have already been removed from the model by now
-    assert(!_model.contains(snapshot));
+    assert(!_model!.contains(snapshot));
     if (mounted) {
       try {
         setState(() {
@@ -239,7 +236,7 @@ class FirestoreAnimatedListState extends State<FirestoreAnimatedList> {
           );
         });
       } catch (error) {
-        _model.log("Failed to remove Widget on index $index");
+        _model!.log("Failed to remove Widget on index $index");
       }
     }
   }
@@ -251,7 +248,7 @@ class FirestoreAnimatedListState extends State<FirestoreAnimatedList> {
     }
   }
 
-  void _onLoaded(QuerySnapshot querySnapshot) {
+  void _onLoaded(QuerySnapshot? querySnapshot) {
     if (mounted && !_loaded) {
       setState(() {
         _loaded = true;
@@ -266,13 +263,16 @@ class FirestoreAnimatedListState extends State<FirestoreAnimatedList> {
 
   Widget _buildItem(
       BuildContext context, int index, Animation<double> animation) {
-    return widget.itemBuilder(context, _model[index], animation, index);
+    return widget.itemBuilder(context, _model![index], animation, index);
   }
 
   @override
   Widget build(BuildContext context) {
-     if (_model.isEmpty) {
-      return _loaded ? (widget.emptyChild ?? Container()) : (widget.defaultChild ?? const Center(child: CircularProgressIndicator()));
+    if (_model!.isEmpty) {
+      return _loaded
+          ? (widget.emptyChild ?? Container())
+          : (widget.defaultChild ??
+              const Center(child: CircularProgressIndicator()));
     }
 
     if (_error != null) {
@@ -282,7 +282,7 @@ class FirestoreAnimatedListState extends State<FirestoreAnimatedList> {
     return AnimatedList(
       key: _animatedListKey,
       itemBuilder: _buildItem,
-      initialItemCount: _model.length,
+      initialItemCount: _model!.length,
       scrollDirection: widget.scrollDirection,
       reverse: widget.reverse,
       controller: widget.controller,
