@@ -12,25 +12,25 @@ import 'firestore_list.dart';
 
 typedef Widget FirestoreAnimatedStaggeredItemBuilder(
   BuildContext context,
-  DocumentSnapshot snapshot,
+  DocumentSnapshot? snapshot,
   Animation<double> animation,
   int index,
 );
 
 typedef StaggeredTile FirestoreStaggeredTileBuilder(
   int index,
-  DocumentSnapshot snapshot,
+  DocumentSnapshot? snapshot,
 );
 
 /// An AnimatedList widget that is bound to a query
 class FirestoreAnimatedStaggered extends StatefulWidget {
   /// Creates a scrolling container that animates items when they are inserted or removed.
   FirestoreAnimatedStaggered({
-    Key key,
-    @required this.query,
-    @required this.itemBuilder,
-    @required this.staggeredTileBuilder,
-    @required this.crossAxisCount,
+    Key? key,
+    required this.query,
+    required this.itemBuilder,
+    required this.staggeredTileBuilder,
+    required this.crossAxisCount,
     this.mainAxisSpacing = 4.0,
     this.crossAxisSpacing = 4.0,
     this.childAspectRatio = 1.0,
@@ -49,20 +49,17 @@ class FirestoreAnimatedStaggered extends StatefulWidget {
     this.shrinkWrap = false,
     this.padding,
     this.duration = const Duration(milliseconds: 300),
-  })  : assert(query != null),
-        assert(itemBuilder != null),
-        assert(staggeredTileBuilder != null),
-        assert(crossAxisCount != null && crossAxisCount > 0),
-        assert(mainAxisSpacing != null && mainAxisSpacing >= 0),
-        assert(crossAxisSpacing != null && crossAxisSpacing >= 0),
-        assert(childAspectRatio != null && childAspectRatio > 0),
+  })  : assert(crossAxisCount > 0),
+        assert(mainAxisSpacing >= 0),
+        assert(crossAxisSpacing >= 0),
+        assert(childAspectRatio > 0),
         super(key: key);
 
   /// A Firestore query to use to populate the animated list
   final Query query;
 
   /// Method that gets called once the stream updates with a new [QuerySnapshot]
-  final Function(QuerySnapshot) onLoaded;
+  final Function(QuerySnapshot)? onLoaded;
 
   /// Signature for a function that creates [StaggeredTile] for a given index and [DocumentSnapshot]
   final FirestoreStaggeredTileBuilder staggeredTileBuilder;
@@ -81,7 +78,7 @@ class FirestoreAnimatedStaggered extends StatefulWidget {
 
   /// Called before any operation with a [DocumentSnapshot];
   /// If it returns `true`, then dismisses that [DocumentSnapshot] from the list
-  final FilterCallback filter;
+  final FilterCallback? filter;
 
   /// This will change `onDocumentAdded` call to `.add` instead of `.insert`,
   /// which might help if your query doesn't care about order changes
@@ -89,15 +86,15 @@ class FirestoreAnimatedStaggered extends StatefulWidget {
 
   /// A widget to display while the query is loading. Defaults to a
   /// centered [CircularProgressIndicator];
-  final Widget defaultChild;
+  final Widget? defaultChild;
 
   /// A widget to display if an error ocurred. Defaults to a
   /// centered [Icon] with `Icons.error` and the error itsef;
-  final Widget errorChild;
+  final Widget? errorChild;
 
   /// A widget to display if the query returns empty. Defaults to a
   /// `Container()`;
-  final Widget emptyChild;
+  final Widget? emptyChild;
 
   /// Called, as needed, to build list item widgets.
   ///
@@ -137,7 +134,7 @@ class FirestoreAnimatedStaggered extends StatefulWidget {
   /// view is scrolled.
   ///
   /// Must be null if [primary] is true.
-  final ScrollController controller;
+  final ScrollController? controller;
 
   /// Whether this is the primary scroll view associated with the parent
   /// [PrimaryScrollController].
@@ -147,7 +144,7 @@ class FirestoreAnimatedStaggered extends StatefulWidget {
   ///
   /// Defaults to true when [scrollDirection] is [Axis.vertical] and
   /// [controller] is null.
-  final bool primary;
+  final bool? primary;
 
   /// How the scroll view should respond to user input.
   ///
@@ -155,7 +152,7 @@ class FirestoreAnimatedStaggered extends StatefulWidget {
   /// user stops dragging the scroll view.
   ///
   /// Defaults to matching platform conventions.
-  final ScrollPhysics physics;
+  final ScrollPhysics? physics;
 
   /// Whether the extent of the scroll view in the [scrollDirection] should be
   /// determined by the contents being viewed.
@@ -174,7 +171,7 @@ class FirestoreAnimatedStaggered extends StatefulWidget {
   final bool shrinkWrap;
 
   /// The amount of space by which to inset the children.
-  final EdgeInsets padding;
+  final EdgeInsets? padding;
 
   /// The duration of the insert and remove animation.
   ///
@@ -190,8 +187,8 @@ class FirestoreAnimatedStaggeredState
     extends State<FirestoreAnimatedStaggered> {
   final GlobalKey<AnimatedStaggeredGridState> _animatedListKey =
       GlobalKey<AnimatedStaggeredGridState>();
-  FirestoreList _model;
-  Exception _error;
+  FirestoreList? _model;
+  Exception? _error;
   bool _loaded = false;
 
   /// Should only be called without setState, inside @override methods here
@@ -227,7 +224,7 @@ class FirestoreAnimatedStaggeredState
   @override
   void dispose() {
     // Cancel the Firebase stream subscriptions
-    _model.clear();
+    _model!.clear();
     super.dispose();
   }
 
@@ -251,13 +248,13 @@ class FirestoreAnimatedStaggeredState
         });
       }
     } catch (error) {
-      _model.log("Failed to run onDocumentAdded");
+      _model!.log("Failed to run onDocumentAdded");
     }
   }
 
   void _onDocumentRemoved(int index, DocumentSnapshot snapshot) {
     // The child should have already been removed from the model by now
-    assert(!_model.contains(snapshot));
+    assert(!_model!.contains(snapshot));
     if (mounted) {
       try {
         setState(() {
@@ -270,7 +267,7 @@ class FirestoreAnimatedStaggeredState
           );
         });
       } catch (error) {
-        _model.log("Failed to remove Widget on index $index");
+        _model!.log("Failed to remove Widget on index $index");
       }
     }
   }
@@ -282,7 +279,7 @@ class FirestoreAnimatedStaggeredState
     }
   }
 
-  void _onLoaded(QuerySnapshot querySnapshot) {
+  void _onLoaded(QuerySnapshot? querySnapshot) {
     if (mounted && !_loaded) {
       setState(() {
         _loaded = true;
@@ -297,13 +294,16 @@ class FirestoreAnimatedStaggeredState
 
   Widget _buildItem(
       BuildContext context, int index, Animation<double> animation) {
-    return widget.itemBuilder(context, _model[index], animation, index);
+    return widget.itemBuilder(context, _model![index], animation, index);
   }
 
   @override
   Widget build(BuildContext context) {
-    if (_model.isEmpty) {
-      return _loaded ? (widget.emptyChild ?? Container()) : (widget.defaultChild ?? const Center(child: CircularProgressIndicator()));
+    if (_model!.isEmpty) {
+      return _loaded
+          ? (widget.emptyChild ?? Container())
+          : (widget.defaultChild ??
+              const Center(child: CircularProgressIndicator()));
     }
 
     if (_error != null) {
@@ -313,12 +313,12 @@ class FirestoreAnimatedStaggeredState
     return AnimatedStaggeredGrid(
       key: _animatedListKey,
       staggeredTileBuilder: (index) =>
-          widget.staggeredTileBuilder?.call(index, _model.elementAt(index)),
+          widget.staggeredTileBuilder.call(index, _model!.elementAt(index)),
       crossAxisCount: widget.crossAxisCount,
       mainAxisSpacing: widget.mainAxisSpacing,
       crossAxisSpacing: widget.crossAxisSpacing,
       itemBuilder: _buildItem,
-      initialItemCount: _model.length,
+      initialItemCount: _model!.length,
       scrollDirection: widget.scrollDirection,
       reverse: widget.reverse,
       controller: widget.controller,
