@@ -10,10 +10,8 @@ abstract class StreamSubscriberMixin<T> {
   List<StreamSubscription<T>> _subscriptions = <StreamSubscription<T>>[];
 
   /// Listens to a stream and saves it to the list of subscriptions.
-  void listen(Stream<T> stream, void onData(T data), {Function onError}) {
-    if (stream != null) {
-      _subscriptions.add(stream.listen(onData, onError: onError));
-    }
+  void listen(Stream<T> stream, void onData(T data), {Function? onError}) {
+    _subscriptions.add(stream.listen(onData, onError: onError));
   }
 
   /// Cancels all streams that were previously added with listen().
